Move Cache.java
