+++ read-api/src/main/java/onlydust/com/marketplace/api/read/ReadApiConfiguration.java
@@ -1,8 +1,8 @@
 package onlydust.com.marketplace.api.read;
 
 import jakarta.persistence.EntityManager;
+import onlydust.com.marketplace.api.read.cache.Cache;
 import onlydust.com.marketplace.api.read.mapper.NotificationMapper;
-import onlydust.com.marketplace.api.read.properties.Cache;
 import onlydust.com.marketplace.api.read.repositories.*;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
