+++ read-api/src/main/java/onlydust/com/marketplace/api/read/adapters/ReadContributionsApiPostgresAdapter.java
@@ -3,8 +3,8 @@ package onlydust.com.marketplace.api.read.adapters;
 import lombok.AllArgsConstructor;
 import onlydust.com.marketplace.api.contract.ReadContributionsApi;
 import onlydust.com.marketplace.api.contract.model.*;
+import onlydust.com.marketplace.api.read.cache.Cache;
 import onlydust.com.marketplace.api.read.entities.bi.ContributionReadEntity;
-import onlydust.com.marketplace.api.read.properties.Cache;
 import onlydust.com.marketplace.api.read.repositories.ContributionReadRepository;
 import onlydust.com.marketplace.api.rest.api.adapter.authentication.AuthenticatedAppUserService;
 import org.springframework.context.annotation.Profile;
@@ -15,8 +15,8 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.*;
 
 import static onlydust.com.marketplace.api.contract.model.ContributionEventEnum.*;
-import static onlydust.com.marketplace.api.read.properties.Cache.XS;
-import static onlydust.com.marketplace.api.read.properties.Cache.ZERO;
+import static onlydust.com.marketplace.api.read.cache.Cache.XS;
+import static onlydust.com.marketplace.api.read.cache.Cache.ZERO;
 import static onlydust.com.marketplace.kernel.exception.OnlyDustException.notFound;
 import static onlydust.com.marketplace.kernel.pagination.PaginationHelper.hasMore;
 import static onlydust.com.marketplace.kernel.pagination.PaginationHelper.nextPageIndex;
