+++ read-api/src/main/java/onlydust/com/marketplace/api/read/adapters/ReadLanguagesApiPostgresAdapter.java
@@ -3,8 +3,8 @@ package onlydust.com.marketplace.api.read.adapters;
 import lombok.AllArgsConstructor;
 import onlydust.com.marketplace.api.contract.ReadLanguagesApi;
 import onlydust.com.marketplace.api.contract.model.LanguagesResponse;
+import onlydust.com.marketplace.api.read.cache.Cache;
 import onlydust.com.marketplace.api.read.entities.LanguageReadEntity;
-import onlydust.com.marketplace.api.read.properties.Cache;
 import onlydust.com.marketplace.api.read.repositories.LanguageReadRepository;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.domain.Sort;
@@ -12,14 +12,14 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RestController;
 
-import static onlydust.com.marketplace.api.read.properties.Cache.M;
+import static onlydust.com.marketplace.api.read.cache.Cache.M;
 
 @RestController
 @AllArgsConstructor
 @Transactional(readOnly = true)
 @Profile("api")
 public class ReadLanguagesApiPostgresAdapter implements ReadLanguagesApi {
-    
+
     private final Cache cache;
     private final LanguageReadRepository languageReadRepository;
