+++ read-api/src/main/java/onlydust/com/marketplace/api/read/adapters/ReadProjectCategoriesApiPostgresAdapter.java
@@ -3,8 +3,8 @@ package onlydust.com.marketplace.api.read.adapters;
 import lombok.AllArgsConstructor;
 import onlydust.com.marketplace.api.contract.ReadProjectCategoriesApi;
 import onlydust.com.marketplace.api.contract.model.ProjectCategoriesResponse;
+import onlydust.com.marketplace.api.read.cache.Cache;
 import onlydust.com.marketplace.api.read.entities.project.ProjectCategoryReadEntity;
-import onlydust.com.marketplace.api.read.properties.Cache;
 import onlydust.com.marketplace.api.read.repositories.ProjectCategoryReadRepository;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.domain.Sort;
@@ -12,7 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RestController;
 
-import static onlydust.com.marketplace.api.read.properties.Cache.M;
+import static onlydust.com.marketplace.api.read.cache.Cache.M;
 
 @RestController
 @AllArgsConstructor
